Refactor exception for TOO_MANY_CONNECTIONS
