+++ shardingsphere-infra/shardingsphere-infra-common/src/main/java/org/apache/shardingsphere/infra/exception/dialect/TooManyConnectionsException.java
@@ -15,12 +15,14 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.proxy.frontend.exception;
+package org.apache.shardingsphere.infra.exception.dialect;
+
+import org.apache.shardingsphere.infra.util.exception.inside.InsideDialectSQLException;
 
 /**
- * Frontend too many connections exception.
+ * Too many connections exception.
  */
-public final class FrontendTooManyConnectionsException extends FrontendException {
+public final class TooManyConnectionsException extends InsideDialectSQLException {
     
     private static final long serialVersionUID = -4397915988239251541L;
 }
